requirements.txt
asgiref==3.8.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django>=4.2.19, <5.0.0
django-cors-headers==4.3.1
django-environ==0.12.0
django-filter==24.2
django-allauth==65.3.1
django-storages==1.14.3
dj-rest-auth==7.0.1
pillow==11.1.0
drf-spectacular==0.28.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
Faker==25.0.1
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
gunicorn==23.0.0
idna==3.7
importlib_metadata==8.4.0
Markdown==3.6
packaging==24.2
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
shortuuid==1.0.13
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.2.1
zipp==3.18.1
django-environ==0.12.0
